skills/q-presentations/requirements.txt
google-genai
Pillow
python-pptx
python-dotenv
